+++ packages/types/src/interfaces.ts
@@ -9,7 +9,7 @@ import {IEvent} from '@polkadot/types/types';
 export interface SubstrateBlock extends SignedBlock {
   // parent block's spec version, can be used to decide the correct metadata that should be used for this block.
   specVersion: number;
-  timestamp: Date;
+  timestamp: Date | undefined;
   events: EventRecord[];
 }